Made changes to mouse position logic and added new drawing tool icon.
Separated the logic for getting the mouse position from converting to
linear index form, to allow for both operations more easily.
Furthermore, some minor changes were made to the conditional checks such
that the index validity is only checked when the user is actively
clicking the left-mouse button over the canvas. Otherwise, added
styling for line tool to the canvas toolbar, and also renamed custom
type for currently active tool.