Implemented line drawing.
Added a tool for drawing lines, using Bresenham's line-drawing
algorithm. In addition, some behavior was added to the canvas so that
dragging across with the line tool first previews the line before
creating it. Moving the mouse around with the line tool while the LMB is
clicked down will then move the line preview around accordingly.