Changed to using mouse-down event listener.
Previously used to listen for a mouse click, which could accidentally
trigger a change in the active glyph if the user was drawing from the
canvas to the area of the glyph set (i.e., keeping the LMB pressed).
Now, the active glyph will change iff the mouse button is explicitly
pressed on a glyph in the glyph set, thereby avoiding this bug and
appearing as a smoother transition.