Added new "outline" tool and made slight adjustments to logic.
A new "outline" tool was added, which if triggered, will toggle every
cell in the canvas to its opposite state; that is, any filled cell will
become empty and vice versa. Thus, if the user draws a glyph in black,
they can view its corresponding outline in white via this tool.
Otherwise, some adjustments were made to the logic; specifically, clean
up was done to the logical checks for the glyph canvas and active glyph
state variables, among other things.