Added guidelines to the canvas.
Note that this also involved bringing back the drawing of gridlines in
the canvas. This subroutine was previously removed, such that the
background instead gave the illusion of there being gridlines, with gaps
between the cells. This was changed so that, when the canvas is given a
decreased opacity for the model glyph in the background, there won't be
gaps in the canvas where the model glyph "bleeds" through.