Added options menu to the toolbar.
This menu contains options to toggle visibility of the model glyph and
the guidelines. A separate component was created to render this, but it
might be best to generalize the "menu" component so it can be applied to
the shapes menu, options, and potentially any future menus (e.g. one for
the fill/invert tools) if included to make room in the toolbar.