Added history tracking for shape tools and actions.
Now, the history is updated accordingly when the user
fills/inverts/clears the
canvas, or draws a shape. Unfortunately, the state management for these
events varies from the draw/erase functionality; the changes made to
account for this might seem a bit "hacky." I'll consider a better
implementation of the history tracking in the future, but for the time
being this achieves the current goal.