Refactored history tracking and implemented redo.
Previously, we had a an array of boolean arrays to keep track of all the
states, treating it as a stack where the latest change gets pushed to
the front of the array, and an undo operation pops the front element.

To allow for redo, we refactor the state management to keep track of two
things for the history: an array of the "states" observed so far, and an
"index" to keep track of the current state on the canvas. If we undo,
this index shifts back by one; and if we redo, the index shifts forward
by one (assuming these operations are feasible). If we make a change to
the canvas, then we take the current states array up to the current
index, then append the new state. (This generalizes to cases where we
make an update after at least one undo, where we overwrite the history
of states following the reverted state and new change.)