Restructured types to incorporate global reducer.
The type system was restructured to explicitly define types of actions;
specifically, actions which impact the main drawing canvas versus the
glyph set. Actions to either of these entities affect the overall "font
state," which is why they act on the same state; however, this
distinction is made to avoid confusion in the future. For instance, it
is important to understand that the "canvas history" is specific to the
drawing canvas, meaning that it will keep track of all changes in the
canvas for the currently active glyph. However, if the user changes the
active glyph, this is an action on the glyph set which should also reset
the canvas history.