Refactored state management to use a reducer.
Previously, the overall state management of the font was handled through
several state variables and their associated setters. This turned out to
be a hassle, since passing state down to the child components would
require several props to be defined. Furthermore, these state variables
were often linked, in the sense that an update to one state variable
could necessitate an update to the other. Thus, all the necessary state
is now maintained in a single object, which we define as the state of
the overall font. Updates are then handled accordingly through a
reducer, which either (conceptually) applies an action to the canvas or
the glyph set. This approach to state management not only results in the
current interface being much cleaner, but also enables the addition of
further actions in the future, as necessary.