Reorganized structure of button and menu components within canvas.
The components for buttons and menus were shifted to a separate file,
due to issues with elements constantly rerendering from being defined
within the Canvas component. To this end, the definitions for these
components were extensively refactored, in particular leveraging new
object types to pass in information about the button type, label, icon,
and so on.