Changed button behavior; some stylistic changes.
The erroneously named "Clear" button was renamed to be a "Reset" button,
with the intention of resetting progress on the current glyph set if the
user wishes to start over. This is to reflect the change made to the
reducer action. Additionally, some stylistic changes were made, as well
as conditions to "disable" certain operations when unnecessary -- e.g.,
disallowing a reset if no progress has been made to begin with.